Port to Sodium 0.5.0

+++ src/main/java/link/infra/indium/renderer/render/SingleBlockLightDataCache.java
@@ -2,7 +2,7 @@ package link.infra.indium.renderer.render;
 
 import java.util.Arrays;
 
-import me.jellysquid.mods.sodium.client.model.light.cache.ArrayLightDataCache;
+import me.jellysquid.mods.sodium.client.model.light.data.ArrayLightDataCache;
 import me.jellysquid.mods.sodium.client.model.light.data.LightDataAccess;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.BlockRenderView;
@@ -17,12 +17,12 @@ public class SingleBlockLightDataCache extends LightDataAccess {
 	private static final int NEIGHBOR_BLOCK_RADIUS = 2;
 	private static final int BLOCK_LENGTH = 1 + (NEIGHBOR_BLOCK_RADIUS * 2);
 
-	private final long[] light;
+	private final int[] light;
 
 	private int xOffset, yOffset, zOffset;
 
 	public SingleBlockLightDataCache() {
-		this.light = new long[BLOCK_LENGTH * BLOCK_LENGTH * BLOCK_LENGTH];
+		this.light = new int[BLOCK_LENGTH * BLOCK_LENGTH * BLOCK_LENGTH];
 	}
 
 	public void reset(BlockPos origin, BlockRenderView blockView) {
@@ -30,7 +30,7 @@ public class SingleBlockLightDataCache extends LightDataAccess {
 		this.yOffset = origin.getY() - NEIGHBOR_BLOCK_RADIUS;
 		this.zOffset = origin.getZ() - NEIGHBOR_BLOCK_RADIUS;
 
-		Arrays.fill(this.light, 0L);
+		Arrays.fill(this.light, 0);
 
 		this.world = blockView;
 	}
@@ -44,10 +44,10 @@ public class SingleBlockLightDataCache extends LightDataAccess {
 	}
 
 	@Override
-	public long get(int x, int y, int z) {
+	public int get(int x, int y, int z) {
 		int l = this.index(x, y, z);
 
-		long word = this.light[l];
+		int word = this.light[l];
 
 		if (word != 0) {
 			return word;
